Create `highlightElement` utility (#538)
* Create `highlightElement` utility

* Update highlight-element.ts
